Expose parameters for IQE tests
Exposing `BONFIRE_COMPONENTS_MAPPING` in basic.yaml pipeline.
This parameter is already defined in the deploy.yaml task
(as a replacement for `BONFIRE_COMPONENT_NAME`).
We need to expose it at pipeline level to make it accessible for the task.

Exposing `IQE_CJI_CLOWDAPP_NAME` which is part of the [deploy-iqe-cji.sh](https://github.com/RedHatInsights/cicd-tools/pull/120/files)
script but it is not accessible. We are exposing it at pipeline and task levels.